+++ test/test_helper.rb
@@ -61,6 +61,13 @@ def stub_api_requests(api_key)
   stub_post("https://api.swiftner.com/organisation/create", api_key)
   stub_put("https://api.swiftner.com/organisation/update/1", api_key)
   stub_delete("https://api.swiftner.com/organisation/delete/1", api_key)
+
+  stub_get("https://api.swiftner.com/channel/get-channels", [{ id: 1, name: "test", type: "audio", space_id: 1 }].to_json, api_key)
+  stub_get("https://api.swiftner.com/channel/is_channel_live?channel_id=1", { "status" => "live" }.to_json, api_key)
+  stub_get("https://api.swiftner.com/channel/get/1", { id: 1, name: "test", type: "audio", space_id: 1 }.to_json, api_key)
+  stub_post("https://api.swiftner.com/channel/create", api_key)
+  stub_put("https://api.swiftner.com/channel/update/1", api_key)
+  stub_delete("https://api.swiftner.com/channel/delete/1", api_key)
 end
 # rubocop:enable Metrics/LineLength, Metrics/ MethodLength